+++ cartagen/utils/lines/simplification/douglas_peucker.py
@@ -1,8 +1,8 @@
 from shapely import simplify
 
-def simplify_douglas_peucker(line, threshold, preserve_topology=True):
+def simplify_douglas_peucker(geometry, threshold, preserve_topology=True):
     """
-    Distance-based line simplification.
+    Simplify a line or polygon using a distance-based selection.
 
     This algorithm was proposed by Ramer :footcite:p:`ramer:1972` and by Douglas and Peucker.
     :footcite:p:`douglas:1973` It is a line filtering algorithm, which means that it
@@ -18,10 +18,10 @@ def simplify_douglas_peucker(line, threshold, preserve_topology=True):
 
     Parameters
     ----------
-    line : LineString
-        The line to simplify.
+    geometry : LineString, MultiLineString, Polygon, MultiPolygon, LinearRing
+        The geometry to simplify.
     threshold : float
-        The distance thresholdto remove the vertex from the line.
+        The distance threshold to remove the vertex from the line.
     preserve_topology : bool, optional
         If set to True, the algorithm will prevent invalid geometries
         from being created (checking for collapses, ring-intersections, etc).
@@ -29,22 +29,26 @@ def simplify_douglas_peucker(line, threshold, preserve_topology=True):
 
     Returns
     -------
-    LineString
+    LineString, MultiLineString, Polygon, MultiPolygon, LinearRing
 
     See Also
     --------
+    simplify_angular :
+        Simplify a line or polygon by removing vertexes with small angles.
     simplify_lang :
-        Look-ahead distance-based line simplification.
+        Simplify a line or polygon using a look-ahead distance-based selection.
     simplify_li_openshaw :
-        Square grid-based line simplification.
+        Simplify a line or a polygon using a regular grid.
     simplify_raposo :
-        Hexagon-based line simplification.
+        Simplify a line or a polygon using an hexagonal tessellation.
     simplify_reumann_witkam :
-        Directional distance-based line simplification.
+        Simplify a line or polygon using a directional distance-based selection.
+    simplify_topographic :
+        Simplify a line or polygon and mimic hand-made cartographic generalization.
     simplify_visvalingam_whyatt :
-        Area-based line simplification.
+        Simplify a line or polygon using an area-based selection.
     simplify_whirlpool :
-        Epsilon-circle based line simplification.
+        Simplify a line or polygon using an epsilon-circle based selection.
 
 
     References
@@ -57,4 +61,4 @@ def simplify_douglas_peucker(line, threshold, preserve_topology=True):
     >>> simplify_douglas_peucker(line, 1.0)
     <LINESTRING (0 0, 2 0, 5 3)>
     """
-    return simplify(line, threshold, preserve_topology=preserve_topology)
+    return simplify(geometry, threshold, preserve_topology=preserve_topology)