+++ docs/code/reference/polygon_elongation.py
@@ -1,0 +1,2 @@
+# Don't delete, this file prevent sphinx from returning an error
+# It is a placeholder for the API reference algorithms that do not have plot demos